+++ docs/changelog.rst
@@ -12,7 +12,7 @@ No changes.
 Jupyterlab extension
 --------------------
 
-No changes.
+* Fix server URL lookup for JupyterLab 2.2.0+;
 
 1.3.4 2020-07-11
 ================